Require distinct ids, not distinct rows
Distinct by id is faster (index scan) and enables us to set up a primary key.

Note: This will drop duplicate rows silently (= the first row with
certain id wins), but in Finland data there is only a couple of cases.
If this becomes a problem, we'll have to figure out what is the exact
meaning of duplicate ids in the data and handle those cases in some
way.