Apply DISTINCT to osm_id only
By default DISTINCT finds unique rows, but it doesn't make sense for
this query, because neither Helsinki or Finland data have any duplicate rows.

On the other hand, finding unique osm_ids makes sense if it is a
reasonable assumption for the data. Helsinki data has no duplicate
osm_ids, but Finland data has.

For now, assume duplicate osm_ids is a special case that we might solve
later if needed. This enables us to use osm_id as a primary key.