Replace roadsnodesinorder query with a better one
WITH ORDINALITY (PostgreSQL 9.4) is faster and simplifies the query.

Multiple tables in FROM is conceptually a cartesian join, which doesn't
really isn't what we want if there is a join condition. The query
optimizer will typically optimize such queries, but a better way is to
use a qualified join with a proper join condition (JOIN xx ON yy).
http://www.postgresql.org/docs/9.4/static/queries-table-expressions.html#QUERIES-FROM

Also add is_endpoint column, which will make later queries much more
efficient. We have all the information here already, so we might as well
mark the endpoints at this point.